+++ rcamelspe.Rcheck/00_pkg_src/rcamelspe/R/globals.R
@@ -1,4 +1,0 @@
-# Global variables for CRAN check compliance
-if (getRversion() >= "2.15.1") {
-  utils::globalVariables(c("gauge_id", ".data"))
-}